requirements.txt
asgiref==3.3.0
Django==3.1.3
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
